requirements.txt
absl-py==2.0.0
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beartype==0.16.4
bitsandbytes==0.41.1
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
classifier-free-guidance-pytorch==0.2.3
cloudpickle==3.0.0
CoLT5-attention==0.10.14
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.6
dill==0.3.7
efficientnet-pytorch==0.7.1
einops==0.7.0
einops-exts==0.0.4
fairscale==0.4.13
Farama-Notifications==0.0.4
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gymnasium==0.29.1
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lion-pytorch==0.1.2
local-attention==1.9.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
open-clip-torch==2.23.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.24.4
psutil==5.9.6
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scipy==1.11.3
sentencepiece==0.1.99
six==1.16.0
stable-baselines3==2.1.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
tiktoken==0.5.1
timm==0.9.8
tokenizers==0.14.1
tokenmonster==1.1.12
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vector-quantize-pytorch==1.9.14
wcwidth==0.2.8
webdataset==0.2.65
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.2
zetascale==0.7.5
